% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zk.R
\name{wkzk_command}
\alias{wkzk_command}
\title{Apply a command JSON and return the resulting response JSON.}
\usage{
wkzk_command(prover, cmd_json)
}
\arguments{
\item{prover}{A prover handle from [wkzk_new()].}

\item{cmd_json}{A command JSON string.}
}
\value{
The response as a JSON string.
}
\description{
Apply a command JSON and return the resulting response JSON.
}
